x86/up: register pinctrl mappings before device driver probing
If a driver for a device in the pinctrl mapping is probed before
the mapping is registered, the pinctrl driver won't be invoked
to configure the pin mux for the device.  This can happen when
drivers are built-in rather than loadable modules.

By using the subsys_initcall() hook instead of module_init(),
we ensure that we register the pinctrl mappings early enough
before the device drivers are probed.

Signed-off-by: Dan O'Donovan <dan@emutex.com>