x86/up: register pinmap config to disable open-drain mode
On the UP board, if the ODEn bit is set on the pad configuration
it seems to impair some functions on the I/O header such as UART, SPI
and even I2C.  This commits disables it for all UP board I/O header
pins by default.

Some cosmetic tidy-up of the pin map code was also included.

Signed-off-by: Dan O'Donovan <dan@emutex.com>