spi: spi-pxa2xx: use cs_gpio if valid
The spi_device structure contains a field named cs_gpio, which
can be used to specify a GPIO pin to use for chip-select
signalling.  This gpio will be used only if valid and if other
supported methods for overriding the chip-select signalling
are not invoked (i.e. control_cs or gpio_cs in controller_data).

Signed-off-by: Dan O'Donovan <dan@emutex.com>