spi: spi-pxa2xx: add GPIO CS lines for UP platform
The PXA2xx SPI controller on Intel X5-Z83xx SoC provides a
single hardware-controlled chip-select signal.  The UP board
platform requires the provision of a 2nd chip-select, using
GPIO, for application and peripheral compatibility.

This patch adds an option to define an arbitrary number of
GPIO-based chip-selects per platform based on DMI ID matching.

There is already some support for this in the SPI core via
device-tree configuration, and this is utilised to some extent
by this patch.  Ultimately, it would be better if we could avail
of this more fully by passing the same information via
ACPI for x86 platforms like the UP board, instead of hard-coding
it into the PXA2xx driver like this, but that is currently
not possible on the UP platform.

Signed-off-by: Dan O'Donovan <dan@emutex.com>