x86/up: register spidev device on UP Board SPI 2-1
Register a spidev device on UP Board SPI bus #2 chip-select 1 to
allow user-space applications to access to SPI devices connected
on the external I/O pin header.

Note that this is a gpio-based chip-select, and registering this
device effectively causes the gpio pin to be reserved by the SPI
driver.  Adding 'up_board.spidev1=0' to the kernel command line
will prevent this device from being registered.

Signed-off-by: Dan O'Donovan <dan@emutex.com>