x86/up: default up_board.spidev1 to false
The patch:
"x86/up: register spidev device on UP Board SPI 2-1"
added a gpio-based SPI chip-select to the external I/O pin header.

However, this means the GPIO was reserved by the up_board module by default,
which is not the case for other function pins. This caused non-obvious issues
for GPIO users.

Signed-off-by: Javier Arteaga <javier@emutex.com>