mmc: sdhci: increase tuning timeout to prevent subsequent commands failing
This is a workaround for an issue observed on Cherry-Trail UP boards
with Kingston M525 eMMC flash chips.  With the existing tuning timeout of
50ms, subsequent write commands would sometimes fail and this would lead
to data loss and application errors.  With the tuning timeout increased
to 150ms, subsequent command failures were no longer seen in the same
tests (despite tuning timeouts still occurring).